+++ src/dataset_loaders.cpp
@@ -266,9 +266,9 @@ bool VlslamDatasetLoader::Grab(int i,
     return true;
 }
 
-std::unordered_map<int, std::array<float, 6>> VlslamDatasetLoader::GrabPointCloud(int i,
+std::unordered_map<int64_t, std::array<double, 6>> VlslamDatasetLoader::GrabPointCloud(int i,
                                                                                   const cv::Mat &img) {
-    std::unordered_map<int, std::array<float, 6>> out;
+    std::unordered_map<int64_t, std::array<double, 6>> out;
     vlslam_pb::Packet *packet_ptr = dataset_.mutable_packets(i);
     for (auto f : packet_ptr->features()) {
         if (f.status() == vlslam_pb::Feature_Status_INSTATE
@@ -281,9 +281,15 @@ std::unordered_map<int, std::array<float, 6>> VlslamDatasetLoader::GrabPointClou
                 color[0] >>= 1;
                 color[1] >>= 1;
                 color[2] >>= 1;
-                out[f.id()] = {f.xw(0), f.xw(1), f.xw(2), color[0], color[1], color[2]};
+                out[f.id()] = {f.xw(0), f.xw(1), f.xw(2),
+                               static_cast<double>(color[0]),
+                               static_cast<double>(color[1]),
+                               static_cast<double>(color[2])};
             } else {
-                out[f.id()] = {f.xw(0), f.xw(1), f.xw(2), color[0], color[1], color[2]};
+                out[f.id()] = {f.xw(0), f.xw(1), f.xw(2),
+                               static_cast<double>(color[0]),
+                               static_cast<double>(color[1]),
+                               static_cast<double>(color[2])};
             }
         }
     }
